Undo version type clobberage. It really is supposed to be Integer
git-svn-id: http://google-web-toolkit.googlecode.com/svn/branches/2.1@8007 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ bikeshed/src/com/google/gwt/sample/expenses/server/domain/Report.java
@@ -116,7 +116,7 @@ public class Report {
 
   @Version
   @Column(name = "version")
-  private Long version;
+  private Integer version;
 
   private Date created;
   
@@ -160,7 +160,7 @@ public class Report {
     return this.reporterKey;
   }
 
-  public Long getVersion() {
+  public Integer getVersion() {
     return this.version;
   }
 
@@ -207,7 +207,7 @@ public class Report {
     this.reporterKey = reporter;
   }
 
-  public void setVersion(Long version) {
+  public void setVersion(Integer version) {
     this.version = version;
   }
